+++ contracts/ElectoralSystems/PluralitySystem.sol
@@ -18,4 +18,32 @@ contract PluralitySystem is ElectoralSystemInterface {
 
         return i;
     }
+
+    function topCandidates(BallotInterface ballot) public view returns (uint[2]) {
+        uint[2] candidates;
+        uint[2] candidateVoteCounts;
+
+        for (uint i = 0; i < ballot.optionsLength(); i++) {
+            uint votes = ballot.votes(i);
+
+            if (votes > candidateVoteCounts[0]) {
+                candidates[1] = candidates[0];
+                candidates[0] = i;
+                candidateVoteCounts[1] = candidateVoteCounts[0];
+                candidateVoteCounts[0] = votes;
+                continue;
+            }
+
+            if (votes > candidateVoteCounts[1] || votes == candidateVoteCounts[0]) {
+                candidates[1] = i;
+                candidateVoteCounts[1] = votes;
+            }
+        }
+
+        return candidates;
+    }
+
+    function hasWinner(BallotInterface) public view returns (bool) {
+        return true; // @todo maybe calc that there are none with the same vote count
+    }
 }