+++ contracts/Proposals/Ballot/Ballot.sol
@@ -19,24 +19,31 @@ contract Ballot is Ownable, BallotInterface {
         uint choice;
     }
 
-    uint public quorum;
-    uint public optionsLength;
+    struct Round {
+        mapping (address => Vote) votes;
+        mapping (uint => Option) options;
 
-    mapping (address => Vote) public votes;
-    mapping (uint => Option) public options;
+        uint optionsLength;
+    }
+
+    uint public quorum;
+    uint public votingRound;
+
+    mapping (uint => Round) rounds;
 
     function Ballot(bytes32[] labels, bytes32[] data, bool[] willAccept) public {
         require(labels.length == data.length && labels.length == willAccept.length);
 
-        optionsLength = labels.length;
+        votingRound = 0;
+        rounds[votingRound].optionsLength = labels.length;
 
-        for (uint i = 0; i < optionsLength; i++) {
+        for (uint i = 0; i < rounds[votingRound].optionsLength; i++) {
             Mode mode = Mode.Accept;
             if (!willAccept[i]) {
                 mode = Mode.Reject;
             }
 
-            options[i] = Option({votes: 0, label: labels[i], data: data[i], mode: mode});
+            rounds[votingRound].options[i] = Option({votes: 0, label: labels[i], data: data[i], mode: mode});
         }
     }
 
@@ -48,9 +55,9 @@ contract Ballot is Ownable, BallotInterface {
         require(weight > 0);
 
         quorum += weight;
-        options[choice].votes += weight;
 
-        votes[voter] = Vote({choice: choice, weight: weight});
+        rounds[votingRound].options[choice].votes += weight;
+        rounds[votingRound].votes[voter] = Vote({choice: choice, weight: weight});
     }
 
     /// @dev Undoes a voters votes for a choice.
@@ -58,32 +65,52 @@ contract Ballot is Ownable, BallotInterface {
     function unvote(address voter) external onlyOwner {
         require(voted(voter));
 
-        Vote storage vote = votes[voter];
+        Vote storage vote = rounds[votingRound].votes[voter];
 
         quorum -= vote.weight;
-        options[vote.choice].votes -= vote.weight;
+        rounds[votingRound].options[vote.choice].votes -= vote.weight;
 
-        delete votes[voter];
+        delete rounds[votingRound].votes[voter];
+    }
+
+    /// @dev Starts a new voting round with choices.
+    /// @param choices Choices included in the next round.
+    function nextRound(uint[2] choices) external onlyOwner {
+
+        uint256 previousVotingRound = votingRound;
+        votingRound = votingRound + 1;
+
+        rounds[votingRound].optionsLength = choices.length;
+
+        for (uint i = 0; i < choices.length; i++) {
+            Option previous = rounds[previousVotingRound].options[choices[i]];
+            rounds[votingRound].options[i] = Option({
+                votes: 0,
+                label: previous.label,
+                data: previous.data,
+                mode: previous.mode
+            });
+        }
     }
 
     function optionWillAccept(uint index) external view returns (bool) {
-        return options[index].mode == Mode.Accept;
+        return rounds[votingRound].options[index].mode == Mode.Accept;
     }
 
     function getLabel(uint index) external view returns (bytes32) {
-        return options[index].label;
+        return rounds[votingRound].options[index].label;
     }
 
     function getData(uint index) external view returns (bytes32) {
-        return options[index].data;
+        return rounds[votingRound].options[index].data;
     }
 
     function isValidChoice(uint index) external view returns (bool) {
-        return optionsLength > index;
+        return rounds[votingRound].optionsLength > index;
     }
 
     function optionsLength() external view returns (uint) {
-        return optionsLength;
+        return rounds[votingRound].optionsLength;
     }
 
     function quorum() external view returns (uint) {
@@ -94,12 +121,12 @@ contract Ballot is Ownable, BallotInterface {
     /// @param choice Selected option.
     /// @return count of votes for option.
     function votes(uint choice) external view returns (uint) {
-        return options[choice].votes;
+        return rounds[votingRound].options[choice].votes;
     }
 
     /// @dev Whether a voter has voted on a specific proposal.
     /// @param voter Address of the voter.
     function voted(address voter) public view returns (bool) {
-        return votes[voter].weight > 0;
+        return rounds[votingRound].votes[voter].weight > 0;
     }
 }