+++ contracts/Proposals/Ballot/BallotInterface.sol
@@ -4,6 +4,7 @@ interface BallotInterface {
 
     function vote(address voter, uint choice, uint weight) external;
     function unvote(address voter) external;
+    function nextRound(uint[2] choices) external;
     function optionWillAccept(uint index) external view returns (bool);
     function getLabel(uint index) external view returns (bytes32);
     function getData(uint index) external view returns (bytes32);